# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
  description: API de Ecommerce
components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  "/api/v1/auth/sign_up":
    post:
      summary: Registers a user
      tags:
      - Auth
      parameters: []
      responses:
        '201':
          description: user created
        '422':
          description: invalid request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: object
                  properties:
                    email:
                      type: string
                      example: user@example.com
                    password:
                      type: string
                      example: password123
                    name:
                      type: string
                      example: John Doe
                  required:
                  - email
                  - password
                  - name
...
  "/api/v1/categories":
    get:
      summary: Lists all categories
      tags:
      - Categories
      parameters:
      - name: view
        in: query
        required: false
        description: "View type (summary para vista resumida):\n * `summary` \n *
          `minimal` \n "
        enum:
        - summary
        - minimal
        schema:
          type: string
      responses:
        '200':
          description: categories found (full view by default)
    post:
      summary: Creates a category
      tags:
      - Categories
      security:
      - bearer_auth: []
      parameters: []
      responses:
        '201':
          description: category created
        '403':
          description: forbidden for non-admin users
        '422':
          description: invalid request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: object
                  properties:
                    name:
                      type: string
                      example: Electronics
                    description:
                      type: string
                      example: Electronic devices and gadgets
                  required:
                  - name
  "/api/v1/categories/{id}":
    parameters:
    - name: id
      in: path
      description: category id
      required: true
      schema:
        type: string
    get:
      summary: Retrieves a category
      tags:
      - Categories
      responses:
        '200':
          description: category found (full view by default)
        '404':
          description: category not found
    patch:
      summary: Updates a category
      tags:
      - Categories
      security:
      - bearer_auth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: category updated with new products
        '403':
          description: forbidden for non-admin users
        '404':
          description: category not found
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: object
                  properties:
                    name:
                      type: string
                    description:
                      type: string
                    product_ids:
                      type: array
                      items:
                        type: integer
    delete:
      summary: Deletes a category
      tags:
      - Categories
      security:
      - bearer_auth: []
      responses:
        '204':
          description: category deleted
        '403':
          description: forbidden for non-admin users
        '404':
          description: category not found
...
  "/api/v1/users":
    get:
      summary: Lists all client users
      tags:
      - Users
      security:
      - bearer_auth: []
      parameters:
      - name: view
        in: query
        required: false
        description: "View type (summary para vista resumida):\n * `summary` \n "
        enum:
        - summary
        schema:
          type: string
      responses:
        '200':
          description: users found (full view by default)
        '403':
          description: forbidden for non-admin users
  "/api/v1/users/{id}":
    parameters:
    - name: id
      in: path
      description: user id
      required: true
      schema:
        type: string
    get:
      summary: Retrieves a user
      tags:
      - Users
      security:
      - bearer_auth: []
      parameters:
      - name: view
        in: query
        required: false
        description: "View type (summary para vista resumida):\n * `summary` \n "
        enum:
        - summary
        schema:
          type: string
      responses:
        '200':
          description: user found (full view by default)
        '404':
          description: user not found
        '401':
          description: unauthorized
    patch:
      summary: Updates a user
      tags:
      - Users
      security:
      - bearer_auth: []
      parameters: []
      responses:
        '200':
          description: user updated
        '403':
          description: forbidden for non-admin users
        '404':
          description: user not found
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: object
                  properties:
                    name:
                      type: string
                      example: Nuevo Nombre
    delete:
      summary: Deletes a user
      tags:
      - Users
      security:
      - bearer_auth: []
      responses:
        '200':
          description: user deleted
        '403':
          description: forbidden for non-admin users
        '404':
          description: user not found
servers:
- url: http://{defaultHost}
  variables:
    defaultHost:
      default: localhost:3000